requirements.txt
# TheSNMC RustDB dependencies
psycopg[binary]
Pillow
